Pull latest from temp pxcore2 stage

+++ examples/pxScene2d/src/rtNode.cpp
@@ -11,6 +11,21 @@
 #include <iostream>
 #include <sstream>
 
+#ifndef WIN32
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wunused-parameter"
+#endif
+
+#include "node.h"
+#include "node_javascript.h"
+
+#include "env.h"
+#include "env-inl.h"
+
+#ifndef WIN32
+#pragma GCC diagnostic pop
+#endif
+
 #include "pxCore.h"
 
 #include "rtNode.h"
@@ -18,8 +33,19 @@
 #include "jsbindings/rtObjectWrapper.h"
 #include "jsbindings/rtFunctionWrapper.h"
 
-#include "node.h"
-#include "node_javascript.h"
+
+#ifdef RUNINMAIN
+
+#include "pxEventLoop.h"
+extern pxEventLoop* gLoop;
+
+#define ENTERSCENELOCK()
+#define EXITSCENELOCK() 
+#else
+#define ENTERSCENELOCK() rtWrapperSceneUpdateEnter();
+#define EXITSCENELOCK()  rtWrapperSceneUpdateExit(); 
+#endif
+
 
 using namespace v8;
 using namespace node;
@@ -43,10 +69,6 @@ static const char** exec_argv;
 static DWORD __rt_main_thread__;
 #else
 static pthread_t __rt_main_thread__;
-static pthread_t __rt_render_thread__;
-static pthread_mutex_t gInitLock = PTHREAD_MUTEX_INITIALIZER;
-static pthread_cond_t gInitCond  = PTHREAD_COND_INITIALIZER;
-static bool gInitComplete = false;
 #endif
 
 bool rtIsMainThread()
@@ -54,7 +76,8 @@ bool rtIsMainThread()
 #ifdef WIN32
   return GetCurrentThreadId() == __rt_main_thread__;
 #else
-  return pthread_self() == __rt_main_thread__;
+//  return pthread_self() == __rt_main_thread__;
+  return pthread_self() != __rt_main_thread__;
 #endif
 }
 
@@ -63,7 +86,7 @@ bool rtIsMainThread()
 
 void *uvThread(void *ptr)
 {
-  printf("uvThread() - ENTER\n");
+  //printf("uvThread() - ENTER\n");
 
   if(ptr)
   {
@@ -71,7 +94,7 @@ void *uvThread(void *ptr)
 
     if(ctx)
     {
-      ctx->uvWorker();
+      ctx->uvWorker(); // BLOCKS
     }
     else
     {
@@ -89,8 +112,6 @@ void *uvThread(void *ptr)
 
 void *jsThread(void *ptr)
 {
-//  printf("jsThread() - ENTER\n");
-
   if(ptr)
   {
     rtNodeContext *ctx = (rtNodeContext*) ptr;
@@ -117,6 +138,16 @@ void *jsThread(void *ptr)
   return NULL;
 }
 
+static void timerCallback(uv_timer_t* )
+{
+    #ifdef RUNINMAIN
+    if (gLoop)
+      gLoop->runOnce();
+    #else
+    rtLogDebug("uv timer callback");
+    #endif
+}
+
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -131,12 +162,19 @@ static inline bool file_exists(const char *file)
 
 
 rtNodeContext::rtNodeContext(v8::Isolate *isolate) :
-   mRefCount(0), mIsolate(isolate), mKillUVWorker(false)
+     mKillUVWorker(false),  js_worker(0), uv_worker(0),
+     mIsolate(isolate), mEnv(NULL), mRefCount(0) //, mTimer() //, mContext(0)
 {
   assert(isolate); // MUST HAVE !
 
   AddRef();
 
+  createEnvironment();
+}
+
+
+void rtNodeContext::createEnvironment()
+{  
   Locker                locker(mIsolate);
   Isolate::Scope isolate_scope(mIsolate);
   HandleScope     handle_scope(mIsolate);
@@ -182,8 +220,36 @@ rtNodeContext::rtNodeContext(v8::Isolate *isolate) :
 
 rtNodeContext::~rtNodeContext()
 {
-  mKillUVWorker = true; // kill UV loop
+  Locker                locker(mIsolate);
+  Isolate::Scope isolate_scope(mIsolate);
+  HandleScope     handle_scope(mIsolate);
+  
+  if(uv_worker)
+  {
+    printf("\nINFO: %s() - kill UV loop", __PRETTY_FUNCTION__); fflush(stdout);
+   
+    mKillUVWorker = true; // kill UV loop
+    
+    (void) pthread_join(uv_worker, NULL);
+  }
+  
+  if(mEnv)
+  {
+    RunAtExit(mEnv);
 
+    mEnv->Dispose();
+    mEnv = NULL;
+  }
+  
+  if(exec_argv)
+  {
+    delete[] exec_argv;
+    exec_argv = NULL;
+    exec_argc = 0;
+  }
+    
+  Release();
+  
   mContext.Reset();
   // NOTE: 'mIsolate' is owned by rtNode.  Don't destroy here !
 }
@@ -206,6 +272,21 @@ void rtNodeContext::add(const char *name, rtValue const& val)
   global->Set(String::NewFromUtf8(mIsolate, name), rt2js(mIsolate, val));
 }
 
+void rtNodeContext::startTimers()
+{
+    rtLogInfo("starting background thread for event loop processing");
+
+    // we start a timer in case there aren't any other evens to the keep the
+    // nodejs event loop alive. Fire a time repeatedly.
+    uv_timer_init(uv_default_loop(), &mTimer);
+    
+    #ifdef RUNINMAIN
+    uv_timer_start(&mTimer, timerCallback, 0, 5);
+    #else
+    uv_timer_start(&mTimer, timerCallback, 1000, 1000);
+    #endif
+}
+
 rtObjectRef rtNodeContext::runThread(const char *file)
 {
   if(file_exists(file) == false)
@@ -215,6 +296,7 @@ rtObjectRef rtNodeContext::runThread(const char *file)
     return rtObjectRef(0);  // ERROR
   }
 
+  startTimers();
   startThread(file);
 
   return rtObjectRef(0);// JUNK
@@ -230,20 +312,14 @@ rtObjectRef rtNodeContext::runScript(const char *script)
     Isolate::Scope isolate_scope(mIsolate);
     HandleScope     handle_scope(mIsolate);    // Create a stack-allocated handle scope.
 
-    // Enter the context for compiling and running ... Persistent > Local
-
     // Get a Local context...
     Local<Context> local_context = node::PersistentToLocal<Context>(mIsolate, mContext);
     Context::Scope context_scope(local_context);
 
     Local<String> source = String::NewFromUtf8(mIsolate, script);
 
-    printf("\nINFO:  <<<<< COMPILE SCRIPT !\n");
-
     // Compile the source code.
     Local<Script> run_script = Script::Compile(source);
-
-    printf("DEBUG:  %15s()    - SCRIPT = %s\n", __FUNCTION__, script);
 
     // Run the script to get the result.
     Local<Value> result = run_script->Run();
@@ -253,9 +329,9 @@ rtObjectRef rtNodeContext::runScript(const char *script)
 
     printf("DEBUG:  %15s()    - RESULT = %s\n", __FUNCTION__, *utf8);  // TODO:  Probably need an actual RESULT return mechanisim
 
-    return rtObjectRef(0);// JUNK
-
   }//scope
+
+    return rtObjectRef(0);// JUNK
 }
 
 int rtNodeContext::startThread(const char *js)
@@ -264,7 +340,7 @@ int rtNodeContext::startThread(const char *js)
 
 //  printf("\n startThread() - ENTER\n");
 
-  if(pthread_create(&worker, NULL, jsThread, (void *) this))
+  if(pthread_create(&js_worker, NULL, jsThread, (void *) this))
   {
     fprintf(stderr, "Error creating thread\n");
     return 1;
@@ -325,7 +401,7 @@ void rtNodeContext::uvWorker()
 {
   bool more;
 
-  printf("\n Start >>> uvWorker() !!!! \n");
+  printf("\n Start >>> %s() !!!! \n", __FUNCTION__);
 
   do
   {
@@ -356,25 +432,32 @@ void rtNodeContext::uvWorker()
 
   } while (more == true && mKillUVWorker == false);
 
-  printf("\n End >>> uvWorker() !!!! \n");
+  printf("\n End >>> %s() !!!! \n", __FUNCTION__);
 
   int code = EmitExit(mEnv);
+  
+  UNUSED_PARAM(code);
+  
   RunAtExit(mEnv);
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-rtNode::rtNode() : mPlatform(NULL) //, mPxNodeExtension(NULL)
+rtNode::rtNode() : mPlatform(NULL)
 {  
+  __rt_main_thread__ = pthread_self();
+  
   nodePath();
 
   mIsolate     = Isolate::New();
   node_isolate = mIsolate; // Must come first !!
 }
 
-rtNode::rtNode(int argc, char** argv) : mPlatform(NULL)//, mPxNodeExtension(NULL)
+rtNode::rtNode(int argc, char** argv) : mPlatform(NULL)
 {
+  __rt_main_thread__ = pthread_self();
+  
   nodePath();
 
   mIsolate     = Isolate::New();
@@ -400,17 +483,17 @@ void rtNode::nodePath()
     {
       ::setenv("NODE_PATH", cwd, 1); // last arg is 'overwrite' ... 0 means DON'T !
 
-      printf("\n\n INFO: NODE_PATH = [%s] << NEW\n", cwd);
+      //printf("\n\n INFO: NODE_PATH = [%s] << NEW\n", cwd);
     }
     else
     {
       printf("\nERROR: failed to set NODE_PATH\n");
     }
   }
-  else
-  {
-     printf("\nINFO:  NODE_PATH =  [%s]  <<<<< ALREADY SET !\n", NODE_PATH);
-  }
+  // else
+  // {
+  //    printf("\nINFO:  NODE_PATH =  [%s]  <<<<< ALREADY SET !\n", NODE_PATH);
+  // }
 }
 
 void rtNode::init(int argc, char** argv)
@@ -436,14 +519,13 @@ void rtNode::term()
 {
   if(node_is_initialized)
   {
-    node_is_initialized = false;
-
-    mIsolate = NULL;
-
     node_isolate->Dispose();
     node_isolate = NULL;
+    mIsolate = NULL;
 
     V8::Dispose();
+
+    node_is_initialized = false; 
 
   //  V8::ShutdownPlatform();
   //  if(mPlatform)
@@ -467,6 +549,8 @@ rtNodeContextRef rtNode::getGlobalContext() const
 
 rtNodeContextRef rtNode::createContext(bool ownThread)
 {
+  UNUSED_PARAM(ownThread);    // not implemented yet.  
+
   rtNodeContextRef ctxref = new rtNodeContext(mIsolate);
 
   ctxref->node = this;
